requirements.txt
flask>=3.0
flask-cors>=4.0
anthropic>=0.30
langchain>=0.2
langchain-community>=0.2
pypdf>=4.0
chromadb>=0.5
sentence-transformers>=3.0
boto3>=1.34
python-dotenv>=1.0
werkzeug>=3.0
